requirements.txt
selenium
pytest
allure-pytest
fixture